requirements.txt
numpy
pandas
torch
torchvision
scikit-learn
pyyaml